requirements.txt
amqp==5.0.9
asgiref==3.5.2
astroid==2.6.6
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-appconf==1.0.5
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==22.1
django-imagekit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
drf-yasg==1.21.4
drf-yasg2==1.19.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
pilkit==2.0
Pillow==8.0.1
platformdirs==2.5.3
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wrapt==1.12.1
